+++ config/application.rb
@@ -10,7 +10,6 @@ module TeSS
   class Application < Rails::Application
     # Initialize configuration defaults for originally generated Rails version.
     config.load_defaults 7.0
-    config.autoload_paths << Rails.root.join('lib/modules')
 
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration can go into files in config/initializers
