+++ lib/ingestors/llm_ingestor.rb
@@ -33,8 +33,8 @@ module Ingestors
       event_page.css('script, link').each { |node| node.remove }
       event_page = event_page.text.squeeze(" \n").squeeze("\n").squeeze("\t").squeeze(' ')
       llm_service_hash = {
-        chatgpt: ChatgptService,
-        willma: WillmaService
+        chatgpt: Llm::ChatgptService,
+        willma: Llm::WillmaService
       }
       llm_service_class = llm_service_hash.fetch(TeSS::Config.llm_scraper['model'].to_sym, nil)
       return unless llm_service_class
