+++ lib/tasks/tess.rake
@@ -141,22 +141,30 @@ namespace :tess do
 
   desc 'run LLM post processing'
   task llm_post_processing: :environment do
+    llm_service_hash = {
+      chatgpt: Llm::ChatgptService,
+      willma: Llm::WillmaService
+    }
+    llm_service_class = llm_service_hash.fetch(TeSS::Config.llm_scraper['model'].to_sym, nil)
+    return unless llm_service_class
+
     prompt = File.read('llm_process_prompt.txt')
     Events.each do |event|
-      needs_processing = event&.llm_object&.needs_processing
-      new_prompt = event&.llm_object&.prompt == prompt
+      needs_processing = event&.llm_interaction&.needs_processing
+      new_prompt = event&.llm_interaction&.prompt == prompt
       future_event = event.end > Time.zone.now
-      unless (needs_processing || new_prompt) && future_event
-        event = GptIngestor.new.post_process_func(event)
-        event.save!
-      end
+      next if (needs_processing || new_prompt) && future_event
+
+      llm_service = llm_service_class.new
+      event = llm_service.post_process_func(event)
+      event.save!
     end
   end
 
   desc 'open all events to being llm processed again'
   task reset_llm_status: :environment do
     Events.where { |event| event.end > Time.zone.now }.each do |event|
-      event&.llm_object&.needs_processing = true
+      event&.llm_interaction&.needs_processing = true
       event.save!
     end
   end
