+++ test/models/event_test.rb
@@ -495,13 +495,13 @@ class EventTest < ActiveSupport::TestCase
     event = Event.new(
       title: 'An event',
       timezone: 'UTC',
-      user: user,
+      user:,
       url: 'https://events.com/1',
       keywords: ['fun times'],
       nodes: [node],
       external_resources_attributes: { '0' => { title: 'test', url: 'https://external-resource.com' } },
       materials: [material],
-      scientific_topic_names: ['Proteins', 'DNA'],
+      scientific_topic_names: %w[Proteins DNA],
       operation_names: ['Variant calling']
     )
 
@@ -524,7 +524,7 @@ class EventTest < ActiveSupport::TestCase
               assert_nil dup.url
               assert_equal [material], dup.materials
               assert_equal [node], dup.nodes
-              assert_equal ['Proteins', 'DNA'], dup.scientific_topic_names
+              assert_equal %w[Proteins DNA], dup.scientific_topic_names
               assert_equal ['Variant calling'], dup.operation_names
               assert_equal 1, dup.external_resources.length
               assert_equal 'test', dup.external_resources.first.title
@@ -672,21 +672,21 @@ class EventTest < ActiveSupport::TestCase
                   'Protein fold prediction', 'Protein fold recognition'], @event.reload.operations_and_synonyms
   end
 
-  test 'can add an llm_object to an event' do
+  test 'can add an llm_interaction to an event' do
     e = events(:scraper_user_event)
-    l = llm_objects(:scrape)
-    e.llm_object = l
+    l = llm_interactions(:scrape)
+    e.llm_interaction = l
     e.save!
-    assert_equal e.llm_object.id, l.id
+    assert_equal e.llm_interaction.id, l.id
     assert_equal l.event_id, e.id
   end
 
-  test 'can destroy an llm_object with an event' do
+  test 'can destroy an llm_interaction with an event' do
     e = events(:scraper_user_event)
-    l = llm_objects(:scrape)
-    e.llm_object = l
+    l = llm_interactions(:scrape)
+    e.llm_interaction = l
     e.save!
-    assert_difference 'LlmObject.count', -1 do
+    assert_difference 'LlmInteraction.count', -1 do
       e.destroy!
     end
   end
